Update to v1.09
By copying the files extracted from the updater (source: www.4players.de) into the install directory. According to the readme, the patch was completed on or around 23 Nov 2005; it was released on 24 Nov:
https://forums.civfanatics.com/threads/patch-1-09-released.142993/
v1.08 was released a day earlier, apparently, and quickly retracted:
https://forums.civfanatics.com/threads/1-08-patch-fixes-black-terrain-issue.142885/#post-3358959
Nowhere to be found on the web, so we'll have to skip over that one.

Apart from copying files, the updater also uses a PakBuild wrapper named Civ4PakSplit.exe to split the 326 MB Art0.fpk into four smaller archives (maximal size 100 MB). Since I'm not putting artwork under source control, the commit doesn't reflect this change.

PakBuild itself was released by Firaxis on or around 1 Nov 2005:
https://forums.civfanatics.com/threads/civ4-pakbuild.136023/

Also not committed: The update includes an installer for Xfire; see this CFC thread:
https://forums.civfanatics.com/threads/what-is-xfire.144038/