Add BtS expansion (v3.01) except GameCore DLL source
Files extracted from the international (MULTI5) DVD version from July 2007, purchased in Germany. v3.0 might be the "US version" that patch 3.2 claims to synchronize with the European versions. A patch v3.01 was, apparently, not released.

To be committed separately: updated GameCore DLL source.

The updaters for the base game and DirectX mentioned in the readme are indeed included on the DVD. The DVD also contains installers for a GameSpy tool named "Comrade" and for SP1 (v8.0) of the MSVC redistributable. The redistributable seems unnecessary; some of those libraries are already present in the BtS install directory, at version 7.1, and I doubt that the game can even work with v8.0. (None of those things are going into this repository.)