Mark XML, Python files changed by patch 3.02
These files are included in patch 3.02, but they're no different from the files on my (European) 3.01 DVD. Maybe they had differed in the US version (v3.00?) mentioned in the v3.02 readme. I've added comments (to be reverted shortly) in order to document the contents of the 3.02 patch in the commit history.