Update to v3.02
Files extracted from the updater (source: 4players.de). According to the readme, the patch was completed on or around 17 July 2007. It was released on 21 July.

Apart from the main executable, these are only the files that differ from my (European) version 3.01. According to insiders, the patch includes some hotfixes not mentioned in the readme file:
https://forums.civfanatics.com/threads/warlords-civ4-patches-available-on-direct2drive-first-bts-patch.232401/
Those bugfixes would have to be in the binaries, or they had already been applied in the European version.