Update to v3.03
Files extracted from the updater (source: 4players.de). According to the readme (only updated in English), the patch was completed on or around 6 Aug 2007. It was released on 13 Aug.

It's not clear what the fix for MS "manifest requirements" is. The MSVC runtimes are the same as in v3.01; maybe it's something in the BtS executable. An updated DirectX 9.0c installer from June 2007 is indeed included in the patch. There's also an undocumented hotfix for the FfH mod (Python crash on Vista):
https://forums.civfanatics.com/threads/bts-patch-version-3-03.238028/page-2#post-5818242
(None of those things belong into this repository.)