GameCore DLL of Solver's patch for BtS 3.03
Solver, a BtS beta tester, had started working on his unofficial patch before v3.03, with an initial release in late July 2007:
http://web.archive.org/web/20071225074730/http://apolyton.net/forums/showthread.php?s=&threadid=168750
This early version is not available to me, indeed, no version of Solver's source code is (though I presume that the source code had been released). Solver's patch was only relevant for some six weeks, so not many mods have included it. (In particular, work on BAI had ended and hadn't started yet on BBAI.)

The compiled DLL is no use; I use it only as a stand-in item for the commit history. I've taken it from Gedemon's Regiments & Ethnic Diversity mod, which makes no DLL changes of its own:
https://forums.civfanatics.com/threads/bts-regiments-ethnic-diversity.238671/page-2#post-5895720

Solver's work on the patch ended on 29 Aug 07, the committed DLL was uploaded by Gedemon on 6 Sept 2007, so it should be the final version.