Update to v3.19
Files extracted from the updater (source: 4players.de). According to the readme, the patch was completed on or around 30 April 2009. It was released on 10 June.

The readme, incorrectly, credits (only) Bhruic. The update to Tectonics v3.16 from Dresden's patch is reverted. Not mentioned in the readme (and not committed here): A no-CD patch is included. I'm guessing that SafeDisc 4 was removed from the BtS executable. The readme seems to omit some other essential changes too:
https://forums.civfanatics.com/threads/civ4-bts-v3-19-patch.323831/#post-8160896