CvDeal static functions from BtS 3.17
Most of the DLL changes made by patch 3.17 don't seem to be present in the Col DLL source. So I'm guessing that Col was forked off an unpublished BtS version in between 3.13 and 3.17, but closer to 3.13. (The bugfixes in 3.17 mostly concern BtS additions like espionage and corporations that aren't relevant for Col.) The CvDeal functions are an easily noticeable example of 3.17 code in Col.